Issue #739: define multi-scale mixin and remove reduction from loss arguments
