+++ deepreg/registry.py
@@ -1,5 +1,5 @@
 from copy import deepcopy
-from typing import Any, Callable, Optional
+from typing import Any, Callable, Dict, Optional
 
 BACKBONE_CLASS = "backbone_class"
 LOSS_CLASS = "loss_class"
@@ -116,7 +116,7 @@ class Registry:
         return decorator
 
     def build_from_config(
-        self, category: str, config: dict, default_args: Optional[dict] = None
+        self, category: str, config: Dict, default_args: Optional[dict] = None
     ) -> Any:
         """
         Build a class instance from config dict.
@@ -171,7 +171,7 @@ class Registry:
         """
         return self.register(category=MODEL_CLASS, name=name, cls=cls, force=force)
 
-    def build_model(self, config: dict, default_args: Optional[dict] = None) -> Any:
+    def build_model(self, config: Dict, default_args: Optional[dict] = None) -> Any:
         """
         Instantiate a registered model class.
 
@@ -196,7 +196,7 @@ class Registry:
         """
         return self.register(category=BACKBONE_CLASS, name=name, cls=cls, force=force)
 
-    def build_backbone(self, config: dict, default_args: Optional[dict] = None) -> Any:
+    def build_backbone(self, config: Dict, default_args: Optional[dict] = None) -> Any:
         """
         Instantiate a registered backbone class.
 
@@ -221,7 +221,7 @@ class Registry:
         """
         return self.register(category=LOSS_CLASS, name=name, cls=cls, force=force)
 
-    def build_loss(self, config: dict, default_args: Optional[dict] = None) -> Callable:
+    def build_loss(self, config: Dict, default_args: Optional[dict] = None) -> Callable:
         """
         Instantiate a registered loss class.
 
@@ -247,7 +247,7 @@ class Registry:
         return self.register(category=METRIC_CLASS, name=name, cls=cls, force=force)
 
     def build_metric(
-        self, config: dict, default_args: Optional[dict] = None
+        self, config: Dict, default_args: Optional[dict] = None
     ) -> Callable:
         """
         Instantiate a registered metric class.
@@ -276,7 +276,7 @@ class Registry:
         )
 
     def build_data_loader(
-        self, config: dict, default_args: Optional[dict] = None
+        self, config: Dict, default_args: Optional[dict] = None
     ) -> Any:
         """
         Instantiate a registered data loader class.
@@ -320,7 +320,7 @@ class Registry:
         )
 
     def build_data_augmentation(
-        self, config: dict, default_args: Optional[dict] = None
+        self, config: Dict, default_args: Optional[dict] = None
     ) -> Callable:
         """
         Instantiate a registered data augmentation class.
